projects/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API server to sit between discord and google authentication"
authors = ["keeb12 <kalebkoebelgd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.0"
SQLAlchemy = "^2.0.23"
google-api-python-client = "^2.111.0"
pytest-mock = "^3.12.0"
google-auth-oauthlib = "^1.2.0"
gunicorn = "^21.2.0"
db = {path = "../../shared/db"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
